chore: rename obsolete scene with required suffix
